requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning and statistics
scikit-learn>=1.0.0

# Topological Data Analysis
giotto-tda>=0.5.0
# Alternative: gudhi>=3.5.0

# Network analysis
networkx>=2.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0  # Optional, for interactive plots

# Dimensionality reduction and clustering
umap-learn>=0.5.0
hdbscan>=0.8.0

# Configuration
pyyaml>=5.4.0

# Neuroimaging (optional, for MRI processing)
# nilearn>=0.9.0
# nibabel>=3.2.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development tools (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0

